Sketch: Fix exception on redraw
===============================

Fixes:
https://forum.freecadweb.org/viewtopic.php?p=458293#p458293

Rationale:
In order to fix B-Spline pole dragging, the order was inverted.
This fixed the B-Spline pole dragging issue, but introduced a
draw before solve approach that is not consistent with the rest
of the Sketcher.

In my parallel development I had already identified this inconsistency,
switched the order, and provided a new mechanism to fix the issue with
the B-Spline pole dragging. This will be merged as part of another PR.

In the meantime, this PR restores the intended behaviour, and let us
identify if the particular reported exception also happens in other
situations.