Enable LGTM to analyze some C++ source code
FreeCAD as a whole is too large for LGTM to analyze the entire codebase, the operation
times out after three hours. By focusing only on the core code, this patch allows the
automated testing framework there to analyze our C++ code, in addition to the Python
and JavaScript analsis it was already doing.