[Draft] Add better error message in get_svg for direction
An AttributeError is raised when `direction=Vector(0,0,0)` and obj is
an Arch::Space on line: https://github.com/FreeCAD/FreeCAD/blob/8ab155b45b26f3a40857ba46e7b7260c57965dc4/src/Mod/Draft/draftfunctions/svg.py#L799

This patch checks if early on if the direction vector and raises a
ValueError with a description of what has gone wrong.

A caveat with this solution is that this new behaviour might break old
code which depends on that invalid directions can be used.